+++ CHANGELOG.md
@@ -8,12 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## \[Unreleased\]
 ### Added
+- Video reading API
+  (<https://github.com/openvinotoolkit/datumaro/pull/521>)
 - Mapillary Vistas dataset format (Import-only)
   (<https://github.com/openvinotoolkit/datumaro/pull/537>)
 - Datumaro can now be installed on Windows on Python 3.9
   (<https://github.com/openvinotoolkit/datumaro/pull/547>)
 - Import for SYNTHIA dataset format
   (<https://github.com/openvinotoolkit/datumaro/pull/532>)
+- Support of `score` attribute in KITTI detetion
+  (<https://github.com/openvinotoolkit/datumaro/pull/571>)
 - Support for Accuracy Checker dataset meta files in formats
   (<https://github.com/openvinotoolkit/datumaro/pull/553>,
   <https://github.com/openvinotoolkit/datumaro/pull/569>)